+++ CliMenu/Base/InputHandler.cs
@@ -1,5 +1,6 @@
 using CliMenu.Helpers;
 using CliMenu.Interfaces;
+using CliMenu.Utils;
 
 namespace CliMenu.Base;
 
@@ -8,11 +9,16 @@ namespace CliMenu.Base;
 /// </summary>
 /// <typeparam name="TComponent">Type of component, constrained to <see cref="BaseComponent{IExecutable, IDisplayer, IInputHandler}"/>.</typeparam>
 public abstract class InputHandler<TComponent>(TComponent component) : Extends<TComponent>(component), IInputHandler
-    where TComponent : BaseComponent<IExecutable, IDisplayer, IInputHandler>
+    where TComponent : BaseComponent<IDisplayer, IInputHandler>
 {
     /// <summary>
-    /// Handles input key presses.
+    /// Event fired when a key is pressed.
     /// </summary>
-    /// <param name="key">The console key pressed.</param>
+    public event EventHandler<SingleArgumentEventArgs<ConsoleKey>>? OnKeyPress;
+
+    /// <inheritdoc/>
+    /// <remarks>
+    /// If you are overriding this method, please call <see cref="OnKeyPress"/> on it. 
+    /// </remarks>
     public abstract void HandleInput(ConsoleKey key);
 }