+++ CliMenu/Helpers/No-Op/NoInput.cs
@@ -1,0 +1,12 @@
+using CliMenu.Interfaces;
+
+namespace CliMenu.Helpers.NoOp;
+
+/// <summary>
+/// A no-op implementation of <see cref="IInputHandler"/> 
+/// </summary>
+public sealed class NoInput : IInputHandler, INoOp
+{
+    /// <inheritdoc/>
+    public void HandleInput(ConsoleKey key) { }
+}