+++ CliMenu/Interfaces/INoOp.cs
@@ -1,0 +1,6 @@
+namespace CliMenu.Interfaces;
+
+/// <summary>
+/// Interface for no-op implementations.
+/// </summary>
+public interface INoOp;